Update I18nPlugin for compatibiity with `@dojo/i18n` alpha.6
Instead of injecting all CLDR data into the build for the given locales,
parse arrays of URLs from calls `@dojo/i18n/cldr/load.default`, load the
CLDR data from those URLs, and inject it into the build. This also
allows `cldr-data` to be removed as a package dependency (although it is
still required for development).