Reduce I18nPlugin memory consumption
Since modules cannot be checked for `@dojo/i18n/cldr/load` usage until
after they have been parsed, their AST objects are stored in a cache
until it can be determined whether to parse them. However, with large
applications this can be a memory hog, so modules are checked as early
as possible to release their AST objects from memory when they are not
needed.